tarefas completas da sprint 3

+++ src/controllers/products.js
@@ -11,7 +11,10 @@ const createProduct = async (req, res) => {
       .returning("*");
 
     if (!product[0] || product.length < 1) {
-      return res.status(400).json({ message: "Erro interno do servidor. Não foi possivel realizar o cadastro, tente novamente" });
+      return res.status(400).json({
+        message:
+          "Erro interno do servidor. Não foi possivel realizar o cadastro, tente novamente",
+      });
     }
 
     return res
@@ -44,6 +47,12 @@ const listProducts = async (req, res) => {
         .join("categorias as c", "produtos.categoria_id", "c.id")
         .select("produtos.*", "c.descricao as categoria_descricao")
         .whereRaw(query, categoria_id);
+
+      if (!products[0]) {
+        return res
+          .status(404)
+          .json({ message: "Não possui produtos com essa categoria" });
+      }
       return res.status(200).json(products);
     }
 
@@ -99,6 +108,18 @@ const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
   try {
+    const pedidosProduto = await knex("pedido_produtos")
+      .where({ produto_id: id })
+      .first();
+
+    if (pedidosProduto) {
+      return res
+        .status(400)
+        .json({
+          message:
+            "Não é possivel deletar o produto, pois existe um pedido atrelado ao mesmo",
+        });
+    }
     const product = await knex("produtos").where({ id }).delete();
     return res.status(204).send();
   } catch (error) {
@@ -106,4 +127,10 @@ const deleteProduct = async (req, res) => {
   }
 };
 
-module.exports = { createProduct, listProducts, updateProduct, detailProduct, deleteProduct };
+module.exports = {
+  createProduct,
+  listProducts,
+  updateProduct,
+  detailProduct,
+  deleteProduct,
+};
